CVS-30307 Version policy

See merge request sph-device-plugin/ovms-c!58

+++ tests/functional/fixtures/server_with_version_policy_fixtures.py
@@ -40,8 +40,7 @@ def start_server_model_ver_policy(request, get_image, get_test_dir,
 
     grpc_port, rest_port = get_ports_for_fixture()
 
-    command = "/ie-serving-py/start_server.sh ie_serving config " \
-              "--config_path /opt/ml/model_ver_policy_config.json " \
+    command = "--config_path /opt/ml/model_ver_policy_config.json " \
               "--port {} --rest_port {}".format(grpc_port, rest_port)
 
     container = client.containers.run(image=get_image, detach=True,
@@ -69,12 +68,21 @@ def model_version_policy_models(get_test_dir,
     face_detection_models = download_two_model_versions
     face_detection_1 = os.path.dirname(face_detection_models[0][0])
     face_detection_1_dir = os.path.join(model_ver_dir, '1')
+    face_detection_1_bin = os.path.join(face_detection_1_dir, 
+                                        os.path.basename(face_detection_models[0][0]))
     face_detection_2 = os.path.dirname(face_detection_models[1][0])
     face_detection_2_dir = os.path.join(model_ver_dir, '2')
+    face_detection_2_bin = os.path.join(face_detection_2_dir, 
+                                        os.path.basename(face_detection_models[1][0]))
     age_gender = os.path.dirname(age_gender_model_downloader[0])
     age_gender_dir = os.path.join(model_ver_dir, '3')
-    if not os.path.exists(model_ver_dir):
-        os.makedirs(model_ver_dir)
+    age_gender_bin = os.path.join(age_gender_dir, 
+                                  os.path.basename(age_gender_model_downloader[0]))
+    if not (os.path.exists(model_ver_dir)
+        and os.path.exists(face_detection_1_bin)
+        and os.path.exists(face_detection_2_bin)
+        and os.path.exists(age_gender_bin)):
+        os.makedirs(model_ver_dir, exist_ok=True)
         copy_tree(face_detection_1, face_detection_1_dir)
         copy_tree(face_detection_2, face_detection_2_dir)
         copy_tree(age_gender, age_gender_dir)
